feat: zod-to-openapi v3 support
Upgrades the @asteasolutions/zod-to-openapi dependency to v3. Requires
manually passing in the version to the generator, and its no longer a
part of their default config object. Creating an additional parameter to
set this which will be a breaking change. Also, needed to read the refId
from deeper in the openapi object in their special Zod type.

BREAKING CHANGE: openApiVersion will need to be passed as a separate
argument.